+++ CHANGES.rst
@@ -6,6 +6,7 @@ Changes
 
 - Per-webmap measurement SRS configuration.
 - Exclusive layers groups on webmaps.
+- Extent information in PostGIS layer diagnostics.
 
 
 4.7.0